Add pagekite_add_ssl_cert_name()
This makes it possible to set the certificate name used for SNI.

Expose this option in pagekitec. Enable SNI when the option is set.
Note that the PK_WITH_FRONTEND_SNI actually doesn't do anything except
when a whitelabel is also given, but for consistency it makes sense to
enable it.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>